--- a/src/wasm-lib/kcl/src/lib.rs
+++ b/src/wasm-lib/kcl/src/lib.rs
@@ -5,7 +5,6 @@ pub mod docs;
 pub mod engine;
 pub mod errors;
 pub mod executor;
-pub mod math_parser;
 pub mod parser;
 pub mod server;
 pub mod std;