## Performance

The math parser benchmark now runs in only 3.37% of its previous time (30x speedup). Here's the criterion data from parsing [math.kcl](https://github.com/KittyCAD/modeling-app/blob/387f7e0912f2845a2d174f17eb033257f1e11d33/src/wasm-lib/tests/executor/inputs/math.kcl#L0-L1):

```
// Main
time:   [54.612 ms 54.651 ms 54.699 ms]   
// This PR
time:   [1.8411 ms 1.8420 ms 1.8430 ms]
change: [-96.630% -96.626% -96.622%] (p = 0.00 < 0.05)
```

## Algorithmic differences

The previous algorithm received as input a vector of KCL syntax tokens. This meant the previous math parser had to identify which tokens were operands and which were operators. For example, the input "a[0] + (b[0] + c)" has to be parsed to figure out that a[0] is an operand, and to convert the flat vec of tokens into a nested structure (by understanding the parentheses around b[0] + c).

This algorithm is a little different. It receives a vec of _operators_ and _operands_ -- not tokens. The [previous parser step in Winnow](https://github.com/KittyCAD/modeling-app/blob/9b4b079c111a22110f5631b6a9534f9a14a71e8a/src/wasm-lib/kcl/src/parser/parser_impl.rs#L1029) already parsed the KCL syntax tokens into operators and operands, so there's no need for the math parser to redo this work. The Winnow parser also started building a tree of binary expressions, and it understands nested brackets. 

So, my new parser has a lot less work to do (because it reuses previous Winnow steps). My parser really just needs to understand precedence and associativity, and convert the flat vec of operator/operands into a binary expression tree. Because some of these operands are already binary expressions, this is kind of combining expression trees into one big tree.

As a result, I use a simplified version of the "shunting yard" algorithm for converting infix to postfix. This algorithm doesn't have to worry about open/close parentheses (those parentheses have already been converted into expression trees by previous Winnow parsing step). This applies precedence and associativity.

Then I run the `evaluate(rpn)` function to convert the postfix tokens into a binary expression tree. You just read operands/operators off the postfix list. You push operands into a stack, and if you get an operator, you pop its left and right subtree from the stack, make a new tree, and push that tree onto the stack. 

Also, my algorithm iterates over a stack rather than using recursion, which explains at least some performance improvement (though not much compared to reusing the winnow operand/operator parsing). It should reduce the risk of stack overflows.